feat: more scratchpad methods
