+++ autonomi/src/client/high_level/vault/mod.rs
@@ -12,16 +12,12 @@ pub mod user_data;
 pub use key::{derive_vault_key, VaultSecretKey};
 pub use user_data::UserData;
 
-use crate::client::data_types::scratchpad::{Scratchpad, ScratchpadError};
+use crate::client::data_types::scratchpad::ScratchpadError;
 use crate::client::payment::PaymentOption;
 use crate::client::quote::CostError;
 use crate::client::{Client, PutError};
 use ant_evm::AttoTokens;
-use ant_networking::{GetRecordCfg, PutRecordCfg, VerificationKind};
-use ant_protocol::storage::{try_serialize_record, DataTypes, RecordKind, RetryStrategy};
 use ant_protocol::Bytes;
-use libp2p::kad::{Quorum, Record};
-use std::collections::HashSet;
 use std::hash::{DefaultHasher, Hash, Hasher};
 use tracing::info;
 
@@ -80,8 +76,6 @@ impl Client {
         secret_key: &VaultSecretKey,
         content_type: VaultContentType,
     ) -> Result<AttoTokens, PutError> {
-        let mut total_cost = AttoTokens::zero();
-
         let (mut scratch, is_new) = self
             .get_or_create_scratchpad(secret_key, content_type)
             .await?;
@@ -90,116 +84,16 @@ impl Client {
         debug_assert!(scratch.is_valid(), "Must be valid after being signed. This is a bug, please report it by opening an issue on our github");
 
         let scratch_address = scratch.network_address();
-        let scratch_key = scratch_address.to_record_key();
 
-        info!("Writing to vault at {scratch_address:?}",);
+        info!("Writing to vault at {scratch_address:?}");
 
-        let record = if is_new {
-            let (receipt, _skipped_payments) = self
-                .pay_for_content_addrs(
-                    DataTypes::Scratchpad.get_index(),
-                    std::iter::once((scratch.xorname(), scratch.payload_size())),
-                    payment_option,
-                )
-                .await
-                .inspect_err(|err| {
-                    error!("Failed to pay for new vault at addr: {scratch_address:?} : {err}");
-                })?;
-
-            let (proof, price) = match receipt.values().next() {
-                Some(proof) => proof,
-                None => return Err(PutError::PaymentUnexpectedlyInvalid(scratch_address)),
-            };
-
-            total_cost = *price;
-
-            Record {
-                key: scratch_key,
-                value: try_serialize_record(
-                    &(proof, scratch),
-                    RecordKind::DataWithPayment(DataTypes::Scratchpad),
-                )
-                .map_err(|_| {
-                    PutError::Serialization(
-                        "Failed to serialize scratchpad with payment".to_string(),
-                    )
-                })?
-                .to_vec(),
-                publisher: None,
-                expires: None,
-            }
+        let total_cost = if is_new {
+            self.scratchpad_create(scratch, payment_option).await?
         } else {
-            Record {
-                key: scratch_key,
-                value: try_serialize_record(&scratch, RecordKind::DataOnly(DataTypes::Scratchpad))
-                    .map_err(|_| {
-                        PutError::Serialization("Failed to serialize scratchpad".to_string())
-                    })?
-                    .to_vec(),
-                publisher: None,
-                expires: None,
-            }
+            self.scratchpad_update(scratch).await?;
+            AttoTokens::zero()
         };
-
-        let put_cfg = PutRecordCfg {
-            put_quorum: Quorum::Majority,
-            retry_strategy: Some(RetryStrategy::Balanced),
-            use_put_record_to: None,
-            verification: Some((
-                VerificationKind::Crdt,
-                GetRecordCfg {
-                    get_quorum: Quorum::Majority,
-                    retry_strategy: None,
-                    target_record: None,
-                    expected_holders: HashSet::new(),
-                },
-            )),
-        };
-
-        debug!("Put record - scratchpad at {scratch_address:?} to the network");
-        self.network
-            .put_record(record, &put_cfg)
-            .await
-            .inspect_err(|err| {
-                error!(
-                    "Failed to put scratchpad {scratch_address:?} to the network with err: {err:?}"
-                )
-            })?;
 
         Ok(total_cost)
-    }
-
-    /// Returns an existing scratchpad or creates a new one if it does not exist.
-    pub async fn get_or_create_scratchpad(
-        &self,
-        secret_key: &VaultSecretKey,
-        content_type: VaultContentType,
-    ) -> Result<(Scratchpad, bool), PutError> {
-        let client_pk = secret_key.public_key();
-
-        let pad_res = self.scratchpad_get(secret_key).await;
-        let mut is_new = true;
-
-        let scratch = if let Ok(existing_data) = pad_res {
-            info!("Scratchpad already exists, returning existing data");
-
-            info!(
-                "scratch already exists, is version {:?}",
-                existing_data.count()
-            );
-
-            is_new = false;
-
-            if existing_data.owner() != &client_pk {
-                return Err(PutError::VaultBadOwner);
-            }
-
-            existing_data
-        } else {
-            trace!("new scratchpad creation");
-            Scratchpad::new(client_pk, content_type)
-        };
-
-        Ok((scratch, is_new))
     }
 }
